SAM_API/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
numpy==2.2.6
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
